refactor(dkg-core): split primitives in modules
