+++ crates/dkg-cli/src/bin/coordinator.rs
@@ -3,7 +3,7 @@ use dkg_cli::coordinator::{
     opts::{Command, Opts},
 };
 
-use dkg_core::primitives::{BundledJustification, BundledResponses, BundledShares};
+use dkg_core::primitives::states::{BundledJustification, BundledResponses, BundledShares};
 
 use gumdrop::Options;
 use std::process;
