+++ crates/dkg-core/src/test_helpers.rs
@@ -1,5 +1,5 @@
 use super::board::BoardPublisher;
-use super::primitives::{BundledJustification, BundledResponses, BundledShares};
+use super::primitives::states::{BundledJustification, BundledResponses, BundledShares};
 use threshold_bls::group::Curve;
 
 /// An in-memory board used for testing
